requirements.txt
# Production dependencies
aiogram>=3.0.0
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
alembic>=1.12.0
pydantic>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0